Mejorando los queries de categoria y unidad

+++ src/controllers/category.controller.ts
@@ -56,16 +56,37 @@ export async function createCategory(req: Request, res: Response) {
     try{
         const category: Category = req.body;
         const conn = await connect();
+
+        await conn.query({
+            sql: 'SELECT * FROM categorys WHERE category_name = ? limit 1',
+            values: category.category_name
+        }, async function(error, categoryDB: Category[]) {
+
+            if(error) {
+                return res.status(500).json({
+                    ok: false,
+                    message: 'Internal Server error'
+                });
+            }
+
+            if(categoryDB[0]) {
+                return res.status(400).json({
+                    ok :false,
+                    message: 'Category name already exists',
+                });
+            }
+            
+            await conn.query('INSERT INTO category SET ?', category);
     
-        await conn.query('INSERT INTO category SET ?', category);
-    
-        return res.status(200).json({
-            ok: true,
-            message: 'Category created',
-            category
+            return res.status(200).json({
+                ok: true,
+                message: 'Category created',
+                category
+            });
+
         });
 
-    }catch(error){
+    }catch(error) {
         console.log(error);
         return res.status(500).json({
             ok: false,
@@ -81,16 +102,36 @@ export async function updateCategory(req: Request, res: Response) {
         const id = req.params.category_id;
         const updateCategory: Category = req.body;
         const conn = await connect();
+
+        await conn.query({
+            sql: 'SELECT * FROM category WHERE category_id = ?',
+            values: id 
+        }, async function(error, categoryDB: Category[]) {
+            if(error) {
+                return res.status(500).json({
+                    ok: false,
+                    message: 'Internal Server error'
+                });
+            }
+
+            if(!categoryDB[0]) {
+                return res.status(400).json({
+                    ok: false,
+                    message: 'The category does not exist'
+                });
+            }
+
+            await conn.query('UPDATE category SET ? WHERE category_id = ?', [updateCategory, id]);
     
-        await conn.query('UPDATE category SET ? WHERE category_id = ?', [updateCategory, id]);
-    
-        return res.status(200).json({
-            ok: true,
-            message: 'Category updated',
-            category: id
+            return res.status(200).json({
+                ok: true,
+                message: 'Category updated',
+                category: id
+            });
+
         });
 
-    }catch(error){
+    }catch(error) {
         console.log(error);
         return res.status(500).json({
             ok: false,
@@ -105,13 +146,36 @@ export async function deleteCategory(req: Request, res: Response) {
     try{
         const id = req.params.category_id;
         const conn = await connect();
-        await conn.query('DELETE FROM category WHERE category_id = ?', [id]);
+
+        await conn.query({
+            sql: 'SELECT * FROM category WHERE category_id = ? limit 1',
+            values: id
+        }, async function(error, categoryDB: Category[]) {
+            if(error) {
+                return res.status(500).json({
+                    ok: false,
+                    message: 'Internal Server error'
+                });
+            }
+
+            if(!categoryDB[0]) {
+                return res.status(400).json({
+                    ok: false,
+                    message: 'The category does not exist'
+                });
+            }
+
+            await conn.query('DELETE FROM category WHERE category_id = ?', [id]);
     
-        return res.json({
-            message: 'Category deleted',
-            category: id
+            return res.json({
+                ok: true,
+                message: 'Category deleted',
+                category: id
+            });    
+
         });
-    }catch(error){
+
+    }catch(error) {
         console.log(error);
         return res.status(500).json({
             ok: false,
