+++ src/controllers/unit.controller.ts
@@ -17,7 +17,7 @@ export async function getUnits(req: Request, res: Response): Promise<Response> {
             Units: units[0]
         });    
 
-    }catch(error){
+    }catch(error) {
         console.log(error);
         return res.status(500).json({
             ok: false,
@@ -41,7 +41,7 @@ export async function getUnit(req: Request, res: Response): Promise<Response> {
             unit: unit[0]
         });
 
-    }catch(error){
+    }catch(error) {
         console.log(error);
         return res.status(500).json({
             ok: false,
@@ -56,16 +56,37 @@ export async function createUnit(req: Request, res: Response) {
     try{
         const unit: Unit = req.body;
         const conn = await connect();
-    
-        await conn.query('INSERT INTO unit SET ?', unit);
-    
-        return res.status(200).json({
-            ok: true,
-            message: 'Unit created',
-            unit
-        });
 
-    }catch(error){
+        await conn.query({
+            sql: 'SELECT * FROM unit WHERE unit_name = ? LIMIT 1',
+            values: unit.unit_name
+        }, async function(error, unitDB: Unit[]) {
+            if(error) {
+                return res.status(500).json({
+                    ok: false,
+                    message: 'Internal Server error'
+                });
+            }
+
+            if(unitDB[0]) {
+                return res.status(400).json({
+                    ok :false,
+                    message: 'Unit name already exists',
+                });
+            }
+
+            await conn.query('INSERT INTO unit SET ?', unit);
+    
+            return res.status(200).json({
+                ok: true,
+                message: 'Unit created',
+                unit
+            });
+
+        });
+    
+ 
+    }catch(error) {
         console.log(error);
         return res.status(500).json({
             ok: false,
@@ -81,16 +102,35 @@ export async function updateUnit(req: Request, res: Response) {
         const id = req.params.unit_id;
         const updateUnit = req.body;
         const conn = await connect();
-    
-        await conn.query('UPDATE unit SET ? WHERE unit_id = ?', [updateUnit, id]);
-    
-        return res.status(200).json({
-            ok: true,
-            message: 'Unit updated',
-            unit: id
-        });
 
-    }catch(error){
+        await conn.query({
+            sql: 'SELECT * FROM unit WHERE unit_id = ? LIMIT 1',
+            values: id 
+        }, async function(error, unitDB: Unit[]) {
+            if(error) {
+                return res.status(500).json({
+                    ok: false,
+                    message: 'Internal Server error'
+                });
+            }
+
+            if(!unitDB[0]) {
+                return res.status(400).json({
+                    ok: false,
+                    message: 'The unit does not exist'
+                });
+            }
+            
+            await conn.query('UPDATE unit SET ? WHERE unit_id = ?', [updateUnit, id]);
+    
+            return res.status(200).json({
+                ok: true,
+                message: 'Unit updated',
+                unit: id
+            });
+        });
+        
+    }catch(error) {
         console.log(error);
         return res.status(500).json({
             ok: false,
@@ -105,13 +145,36 @@ export async function deleteUnit(req: Request, res: Response) {
     try{
         const id = req.params.unit_id;
         const conn = await connect();
-        await conn.query('DELETE FROM unit WHERE unit_id = ?', [id]);
+
+        conn.query({
+            sql: 'SELECT * FROM unit WHERE unit_id = ? LIMIT 1',
+            values: id
+        }, async function(error, unitDB: Unit[]) {
+            if(error) {
+                return res.status(500).json({
+                    ok: false,
+                    message: 'Internal Server error'
+                });
+            }
+
+            if(!unitDB[0]) {
+                return res.status(400).json({
+                    ok: false,
+                    message: 'The unit does not exist'
+                });
+            }
+
+            await conn.query('DELETE FROM unit WHERE unit_id = ?', [id]);
     
-        return res.json({
-            message: 'Unit deleted',
-            unit: id
+            return res.json({
+                ok: true,
+                message: 'Unit deleted',
+                unit: id
+            });
+
         });
-    }catch(error){
+
+    }catch(error) {
         console.log(error);
         return res.status(500).json({
             ok: false,
